move log monitor to benefit_claims folder
